[GNN] Tut 2: Node Classification with GNNs
Initial commit for the tutorial code for performing node classification
using GNNs.
The dataset used is the Planetoid Cora dataset.
This commit adds code to successfully download the dataset into the
data/ folder to avoid re-downloads every time.
It also adds a .gitignore to ignore the downloaded data files.
It also adds a README file to display the code output as well as explain
a few aspects of the code and in future commits, will also make the
output scatter plots availble for view in the README